refactor: your ship component into its own file
