+++ packages/client/src/utils/directions.ts
@@ -1,6 +1,7 @@
 import { MoveCard, SailPositions, Wind } from "../constants";
 import { Coord, random } from "@latticexyz/utils";
 import { getPositionByVector } from "./trig";
+import { Arrows } from "../layers/phaser/constants";
 
 /**
  * @param coord Initial coordinate
@@ -112,4 +113,22 @@ export function rotationToDirectionName(rotation: number): string {
   if (rotation > 68 && rotation < 112) return (ret += "SE");
 
   return ret;
+}
+
+export function arrowImg(rotation: number) {
+  return rotation == 360 || rotation == 0
+    ? Arrows.Straight
+    : rotation > 270
+    ? Arrows.SoftLeft
+    : rotation == 270
+    ? Arrows.Left
+    : rotation > 180
+    ? Arrows.HardLeft
+    : rotation == 180
+    ? Arrows.UTurn
+    : rotation > 90
+    ? Arrows.HardRight
+    : rotation == 90
+    ? Arrows.Right
+    : Arrows.SoftRight;
 }
